+++ web/projects/portal/src/app/embed/crosstab/embed-crosstab.component.ts
@@ -168,27 +168,29 @@ export class EmbedCrosstabComponent extends CommandProcessor implements OnInit, 
          this.inputRuntimeId = result.posParams?.runtimeId?.path;
          this.queryParams = tree.queryParams;
 
-         (window.inetsoftConnected as BehaviorSubject<boolean>).subscribe((connected) => {
-            if(!this.connected && connected) {
-               this.connected = true;
+         this.subscriptions.add(
+            (window.inetsoftConnected as BehaviorSubject<boolean>).subscribe((connected) => {
+               if(!this.connected && connected) {
+                  this.connected = true;
 
-               if(!!this.errorTimeout) {
-                  clearTimeout(this.errorTimeout);
+                  if(!!this.errorTimeout) {
+                     clearTimeout(this.errorTimeout);
+                  }
+
+                  this.showError = false;
+                  this.openViewsheet();
+                  this.cdRef.detectChanges();
                }
 
-               this.showError = false;
-               this.openViewsheet();
-               this.cdRef.detectChanges();
-            }
-
-            if(!this.connected && !connected) {
-               this.errorTimeout = setTimeout(() => {
-                  this.showError = true;
-                  console.error("InetSoft client not connected. Please make sure to login first.");
-                  this.cdRef.detectChanges();
-               }, 1000);
-            }
-         });
+               if(!this.connected && !connected) {
+                  this.errorTimeout = setTimeout(() => {
+                     this.showError = true;
+                     console.error("InetSoft client not connected. Please make sure to login first.");
+                     this.cdRef.detectChanges();
+                  }, 1000);
+               }
+            })
+         );
       }
       else {
          if(document.body.className.indexOf("app-loaded") == -1) {
@@ -288,10 +290,8 @@ export class EmbedCrosstabComponent extends CommandProcessor implements OnInit, 
       this.vsObject.active = true;
       this.updateVSInfo();
 
-      if(this.vsObject) {
-         this.vsObjectActions = new EmbedCrosstabActions(this.vsObject, null,
-            this.contextProvider, false, null, null, this.miniToolbarService);
-      }
+      this.vsObjectActions = new EmbedCrosstabActions(this.vsObject, this.contextProvider,
+         false, null, null, null, this.miniToolbarService);
    }
 
    // noinspection JSUnusedGlobalSymbols
@@ -308,8 +308,8 @@ export class EmbedCrosstabComponent extends CommandProcessor implements OnInit, 
       }
 
       this.vsObject.active = true;
-      this.vsObjectActions = new EmbedCrosstabActions(this.vsObject, null,
-         this.contextProvider, false, null, null, this.miniToolbarService);
+      this.vsObjectActions = new EmbedCrosstabActions(this.vsObject, this.contextProvider,
+         false, null, null, null, this.miniToolbarService);
    }
 
    // noinspection JSUnusedGlobalSymbols
@@ -473,6 +473,10 @@ export class EmbedCrosstabComponent extends CommandProcessor implements OnInit, 
    }
 
    onOpenContextMenu(event: MouseEvent) {
+      if(!this.vsObjectActions) {
+         return;
+      }
+
       let actions: AssemblyActionGroup[];
 
       if(event.type === "click") {
